refactor: move updateVariable() into VariableService
